+++ .gitignore
@@ -37,6 +37,7 @@ example/
 .pub-cache/
 .pub/
 /build/
+**/analysis_options.yaml
 
 # Web related
 lib/generated_plugin_registrant.dart
