+++ scripts/common/helper.js
