+++ scripts/sync.js
@@ -2,7 +2,7 @@ const WebSocket = require('ws')
 const getPort = require('get-port')
 const inquirer = require('inquirer')
 const { applyPatch } = require('rfc6902')
-const { error, msg, log } = require('./FN')
+const { error, msg, log } = require('./common/FN')
 
 let initData
 let port
@@ -18,10 +18,10 @@ function renderView(cache = true, useRemote = false) {
 }
 
 const tempHandleMap = {
-	web: () => require('./gen_web').initData,
-	pcweb: () => require('./gen_web').initData,
-	mp: () => require('./gen_mp').initData,
-	flutter: () => require('./gen_flutter').initData,
+	web: () => require('./web/gen_web').initData,
+	pcweb: () => require('./web/gen_web').initData,
+	mp: () => require('./mp/gen_mp').initData,
+	flutter: () => require('./flutter/gen_flutter').initData,
 }
 
 async function main(conf) {
